Create Patient_final_cohort_info table
Query to create glucose_final_cohort_info table located in the Team_5 dataset hosted in the sccm-datathon project.
SQL syntax to be runned on BigQuery (Google Cloud)

It contains the cohort information for the patients that have at least 2 or more glucose reading per day. It is derived from the Patients_info table.